requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
Werkzeug==2.2.3
yarl==1.8.2
